将Task5Mapper和Task5Reducer改回了之前Double的版本，集群测试结果有12个标签

+++ src/main/java/com/nju/mplab14/task5/Task5Reducer.java
@@ -11,13 +11,13 @@ public class Task5Reducer extends Reducer<Text, Text, Text, Text> {
 	protected void reduce(Text key, Iterable<Text> values, Context context)
 			throws IOException, InterruptedException{
 		//计数各label的总权重
-		Map<String,Integer> labelMap = new HashMap<>();
+		Map<String,Double> labelMap = new HashMap<>();
 		for(Text value:values){
 			String []split = value.toString().split("\t");
 			String label = split[0];
-			Integer weight = Integer.valueOf(split[1]);
+			Double weight = Double.valueOf(split[1]);
 			if(labelMap.containsKey(label)){
-				Integer oldValue = labelMap.get(label);
+				Double oldValue = labelMap.get(label);
 				labelMap.put(label, oldValue+weight);
 			} else {
 				labelMap.put(label, weight);
@@ -25,8 +25,8 @@ public class Task5Reducer extends Reducer<Text, Text, Text, Text> {
 		}
 
 		//找到权重最大的几个label，随机选一个（task3中可以保证这里的key的values至少有一个）
-		List<Map.Entry<String,Integer>> maxEntrys = new ArrayList<>();
-		for(Map.Entry<String,Integer> entry:labelMap.entrySet()){
+		List<Map.Entry<String,Double>> maxEntrys = new ArrayList<>();
+		for(Map.Entry<String,Double> entry:labelMap.entrySet()){
 			if(maxEntrys.isEmpty() || entry.getValue().equals(maxEntrys.get(0).getValue())){
 				maxEntrys.add(entry);
 			}else if(entry.getValue()>maxEntrys.get(0).getValue()){
